Change the TestApp to just TestApp
