+++ TestApp/Easy/GetSample.cs
@@ -1,12 +1,12 @@
 ﻿using System;
 using System.Text;
+using CurlThin;
 using CurlThin.Enums;
 using CurlThin.Helpers;
-using CurlThin.SafeHandles;
 
-namespace CurlThin.Samples.Easy
+namespace TestApp.Easy
 {
-    internal class HttpHeadersSample : ISample
+    internal class GetSample : ISample
     {
         public void Run()
         {
@@ -19,21 +19,10 @@ namespace CurlThin.Samples.Easy
             {
                 var dataCopier = new DataCallbackCopier();
 
-                CurlNative.Easy.SetOpt(easy, CURLoption.URL, "http://httpbin.org/headers");
+                CurlNative.Easy.SetOpt(easy, CURLoption.URL, "http://httpbin.org/ip");
                 CurlNative.Easy.SetOpt(easy, CURLoption.WRITEFUNCTION, dataCopier.DataHandler);
 
-                // Initialize HTTP header list with first value.
-                var headers = CurlNative.Slist.Append(SafeSlistHandle.Null, "X-Foo: Bar");
-                // Add one more value to existing HTTP header list.
-                CurlNative.Slist.Append(headers, "X-Qwerty: Asdfgh");
-
-                // Configure libcurl easy handle to send HTTP headers we configured.
-                CurlNative.Easy.SetOpt(easy, CURLoption.HTTPHEADER, headers.DangerousGetHandle());
-
                 var result = CurlNative.Easy.Perform(easy);
-
-                // Cleanup HTTP header list after request has complete.
-                CurlNative.Slist.FreeAll(headers);
 
                 Console.WriteLine($"Result code: {result}.");
                 Console.WriteLine();
